backend/thumbnail-generator/requirements.txt
boto3==1.34.34
Pillow==10.2.0
